Move SparqlItem to item/ and remove qid from Item.
