+++ tests/test_sparql_item.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
 
 from src import console
-from src.models.wikimedia.wikidata.sparql_item import SparqlItem, Value
+from src.models.wikimedia.wikidata.item.sparql import SparqlItem, Value
 
 
 class TestSparqlItem(TestCase):
